pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-jwt>=1.11.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=1.7.1",
    "python-dateutil>=2.9.0.post0",
    "python-decouple>=3.8",
    "requests>=2.32.4",
]
